Create prototype
This is a prototype of the overall project.  It includes most of the
visual elements, some configuration items and much of the GPX file
creation mechanism.  It does not yet show the correct value for distance
(which is currently an artificial value that increments by time) and
the settings page does not yet have any effect.

Signed-off-by: Ed Beroset <beroset@ieee.org>